test: move teste de professor para a pasta de teste
